MatrixBandwidth

Documentation for MatrixBandwidth.

MatrixBandwidth.Exact

MatrixBandwidth.Heuristic

MatrixBandwidth.MatrixBandwidth

MatrixBandwidth.Metaheuristic

MatrixBandwidth.AbstractSolver

MatrixBandwidth.BandwidthResult

MatrixBandwidth.Exact.ExactSolver

MatrixBandwidth.Exact.MBID

MatrixBandwidth.Exact.MBPS

MatrixBandwidth.Heuristic.CuthillMcKee

MatrixBandwidth.Heuristic.HeuristicSolver

MatrixBandwidth.Heuristic.ReverseCuthillMcKee

MatrixBandwidth.Metaheuristic.GRASP

MatrixBandwidth.Metaheuristic.GeneticAlgorithm

MatrixBandwidth.Metaheuristic.MetaheuristicSolver

MatrixBandwidth.Metaheuristic.SimulatedAnnealing

MatrixBandwidth.NotImplementedError

MatrixBandwidth.Heuristic._assert_valid_node_selector

MatrixBandwidth.Heuristic._connected_components

MatrixBandwidth.Heuristic._connected_cuthill_mckee_ordering

MatrixBandwidth.Heuristic.pseudo_peripheral_node

MatrixBandwidth._sym_minimal_band_ordering

MatrixBandwidth.approach

MatrixBandwidth.minimize_bandwidth

MatrixBandwidth.MatrixBandwidth — Module

MatrixBandwidth

Exact, heuristic, and metaheuristic algorithms for matrix bandwidth minimization in Julia.

Full documentation is available for the latest development version of this package.

MatrixBandwidth.AbstractSolver — Type

AbstractSolver

TODO: Write here

MatrixBandwidth.BandwidthResult — Type

BandwidthResult

TODO: Write here

MatrixBandwidth.NotImplementedError — Type

NotImplementedError(f, arg, subtype, abstracttype)

An exception indicating that a function lacks dispatch to handle a specific argument type.

Semantically, this differs from MethodError in that it connotes a developer-side failure to implement a method rather than erroneous user input. Throughout this package, it is often used to warn when an existing function with multiple dispatch on some abstract type is called on a newly created subtype for which no method has been defined.

Fields

f::Function: the function called.

arg::Symbol: the name of the argument with the unsupported type.

subtype::Type: the type of the argument. May be the actual concrete type or some intermediate supertype. (For instance, if the relevant input has concrete type A with hierarchy A <: B <: C and the abstracttype field is C, then both A and B are perfectly valid choices for subtype.)

abstracttype::Type: the abstract type under which the argument is meant to fall.

Constructors

NotImplementedError(::Function, ::Symbol, ::Type, ::Type): constructs a new NotImplementedError instance. Throws an error if the second type is not abstract or the first type is not a subtype of the second.

MatrixBandwidth._sym_minimal_band_ordering — Method

_sym_minimal_bandwidh_ordering(A::AbstractMatrix{Bool}, solver::AbstractSolver)
    -> Vector{Int}

TODO: Write here

MatrixBandwidth.approach — Method

approach(solver::AbstractSolver) -> Symbol

TODO: Write here

MatrixBandwidth.minimize_bandwidth — Method

minimize_bandwidth(A, solver=ReverseCuthillMcKee()) -> BandwidthResult

TODO: Write here

MatrixBandwidth.Exact — Module

MatrixBandwidth.Exact

Exact solvers for matrix bandwidth minimization.

This submodule is part of the MatrixBandwidth.jl package.

MatrixBandwidth.Exact.ExactSolver — Type

ExactSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Exact.MBID — Type

MBID <: ExactSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Exact.MBPS — Type

MBPS <: ExactSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Heuristic — Module

MatrixBandwidth.Heuristic

Heuristic solvers for matrix bandwidth minimization.

This submodule is part of the MatrixBandwidth.jl package.

MatrixBandwidth.Heuristic.CuthillMcKee — Type

CuthillMcKee <: HeuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Heuristic.HeuristicSolver — Type

HeuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Heuristic.ReverseCuthillMcKee — Type

ReverseCuthillMcKee <: HeuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Heuristic._assert_valid_node_selector — Method

_assert_valid_node_selector(selector) -> Nothing

TODO: Write here

MatrixBandwidth.Heuristic._connected_components — Method

_connected_components(A) -> Vector{Vector{Int}}

TODO: Write here

MatrixBandwidth.Heuristic._connected_cuthill_mckee_ordering — Method

_connected_cuthill_mckee_ordering(A, node_selector) -> Vector{Int}

TODO: Write here

MatrixBandwidth.Heuristic.pseudo_peripheral_node — Method

pseudo_peripheral_node(A::AbstractMatrix{Bool}) -> Int

TODO: Write here

MatrixBandwidth.Metaheuristic — Module

MatrixBandwidth.Metaheuristic

Metaheuristic solvers for matrix bandwidth minimization.

This submodule is part of the MatrixBandwidth.jl package.

MatrixBandwidth.Metaheuristic.GRASP — Type

GRASP <: MetaheuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Metaheuristic.GeneticAlgorithm — Type

GeneticAlgorithm <: MetaheuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Metaheuristic.MetaheuristicSolver — Type

MetaheuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Metaheuristic.SimulatedAnnealing — Type

SimulatedAnnealing <: MetaheuristicSolver <: AbstractSolver

TODO: Write here