MatrixBandwidth

Documentation for MatrixBandwidth.

MatrixBandwidth.MatrixBandwidth

MatrixBandwidth.AbstractSolver

MatrixBandwidth.BandwidthResult

MatrixBandwidth.CuthillMcKee

MatrixBandwidth.ExactSolver

MatrixBandwidth.GRASP

MatrixBandwidth.GeneticAlgorithm

MatrixBandwidth.HeuristicSolver

MatrixBandwidth.MBID

MatrixBandwidth.MBPS

MatrixBandwidth.MetaheuristicSolver

MatrixBandwidth.ReverseCuthillMcKee

MatrixBandwidth.SimulatedAnnealing

MatrixBandwidth.approach

MatrixBandwidth.minimize_bandwidth

MatrixBandwidth.MatrixBandwidth — Module

MatrixBandwidth

Exact, heuristic, and metaheuristic algorithms for matrix bandwidth minimization in Julia.

Full documentation is available for the latest development version of this package.

MatrixBandwidth.AbstractSolver — Type

AbstractSolver

TODO: Write here

MatrixBandwidth.BandwidthResult — Type

BandwidthResult

TODO: Write here

MatrixBandwidth.CuthillMcKee — Type

CuthillMcKee <: HeuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.ExactSolver — Type

ExactSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.GRASP — Type

GRASP <: MetaheuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.GeneticAlgorithm — Type

GeneticAlgorithm <: MetaheuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.HeuristicSolver — Type

HeuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.MBID — Type

MBID <: ExactSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.MBPS — Type

MBPS <: ExactSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.MetaheuristicSolver — Type

MetaheuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.ReverseCuthillMcKee — Type

ReverseCuthillMcKee <: HeuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.SimulatedAnnealing — Type

SimulatedAnnealing <: MetaheuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.approach — Method

approach(::AbstractSolver) -> Symbol

TODO: Write here

MatrixBandwidth.minimize_bandwidth — Function

minimize_bandwidth(A, solver=ReverseCuthillMcKee()) -> BandwidthResult

TODO: Write here