MatrixBandwidth.jl – Private API

Documentation for MatrixBandwidth's private API.

Note

The following documentation covers only the private API of the package. For public details, see the public API documentation.

MatrixBandwidth.AbstractSolver — Type

AbstractSolver

TODO: Write here

MatrixBandwidth.NotImplementedError — Type

NotImplementedError(f, arg, subtype, abstracttype)

An exception indicating that a function lacks dispatch to handle a specific argument type.

Semantically, this differs from MethodError in that it connotes a developer-side failure to implement a method rather than erroneous user input. Throughout this package, it is often used to warn when an existing function with multiple dispatch on some abstract type is called on a newly created subtype for which no method has been defined.

Fields

f::Function: the function called.

arg::Symbol: the name of the argument with the unsupported type.

subtype::Type: the type of the argument. May be the actual concrete type or some intermediate supertype. (For instance, if the relevant input has concrete type A with hierarchy A <: B <: C and the abstracttype field is C, then both A and B are perfectly valid choices for subtype.)

abstracttype::Type: the abstract type under which the argument is meant to fall.

Constructors

NotImplementedError(::Function, ::Symbol, ::Type, ::Type): constructs a new NotImplementedError instance. Throws an error if the second type is not abstract or the first type is not a subtype of the second.

MatrixBandwidth.random_sparse_banded_matrix — Method

random_sparse_banded_matrix(n, k; p=0.75, rng=default_rng()) -> Matrix{Float64}

TODO: Write here

MatrixBandwidth.Exact.ExactSolver — Type

ExactSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Heuristic.HeuristicSolver — Type

HeuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Heuristic.pseudo_peripheral_node — Method

pseudo_peripheral_node(A::AbstractMatrix{Bool}) -> Int

Select a pseudo-peripheral node from the connected graph represented by A.

This function acts as a node selector for the Cuthill–McKee and Reverse Cuthill–McKee algorithms, heuristically choosing the node "farthest" from the others in the graph. It is assumed that A is the adjacency matrix of some connected, undirected graph; otherwise, undefined behavior may arise.

Arguments

A::AbstractMatrix{Bool}: the adjacency matrix of some connected, undirected graph. In practice, this semantically represents the connected component of some larger graph.

Returns

Int: the index of the pseudo-peripheral node selected from the graph.

Notes

This function takes heavy inspiration from the implementation in [Net25], which accepts a graph object as input and leverages several pre-existing functions in the networkx library. We herein repurpose the logic to work directly on adjacency matrices, avoiding reallocation overhead and an unnecessary dependency on Graphs.jl.

MatrixBandwidth.Metaheuristic.MetaheuristicSolver — Type

MetaheuristicSolver <: AbstractSolver

TODO: Write here

<!– TODO: Insert a References section here once citations are added –>