MatrixBandwidth.jl – Private API

Documentation for MatrixBandwidth's private API.

Note

The following documentation covers only the private API of the package. For public details, see the public API documentation.

MatrixBandwidth.AbstractAlgorithm — Type

AbstractAlgorithm

Abstract base type for all matrix bandwidth minimization and recognition algorithms.

Interface

Concrete subtypes of AbstractAlgorithm must implement the following methods:

Base.summary(::T) where {T<:AbstractAlgorithm}: returns a String indicating the name of the algorithm (e.g., "Reverse Cuthill–McKee").

_requires_symmetry(::T) where {T<:AbstractAlgorithm}: returns a Bool indicating whether the algorithm requires the input matrix to be structurally symmetric.

Direct subtypes of AbstractAlgorithm must implement the following method:

_problem(::T) where {T<:AbstractAlgorithm}: returns a Symbol indicating the matrix bandwidth problem tackled by the algorithm (e.g., :minimization).

MatrixBandwidth.AbstractResult — Type

AbstractResult

Abstract base type for all matrix bandwidth problem results.

Interface

Concrete subtypes of AbstractResult must implement parametric types

A<:AbstractAlgorithm;

M<:AbstractMatrix{<:Number}; and

O<:Union{Nothing,Vector{Int}},

alongside the following fields:

algorithm::A: the algorithm used to investigate the bandwidth.

matrix::M: the matrix whose bandwidth is investigated.

ordering::O: the corresponding ordering of the rows and columns, if a relevant one is found; otherwise, nothing.

MatrixBandwidth.NotImplementedError — Type

NotImplementedError{Nothing}(f, subtype, abstracttype)
NotImplementedError{Symbol}(f, arg, subtype, abstracttype)

An exception indicating that a function lacks dispatch to handle a specific argument type.

Semantically, this differs from MethodError in that it connotes a developer-side failure to implement a method rather than erroneous user input. Throughout this package, it is often used to warn when an existing function with multiple dispatch on some abstract type is called on a newly created subtype for which no method has been defined.

Fields

f::Function: the function called.

arg::Symbol: the name of the argument with the unsupported type, if the function has multiple arguments. If the function has only one argument, this field should be set to nothing.

subtype::Type: the type of the argument. May be the actual concrete type or some intermediate supertype. (For instance, if the relevant input has concrete type A with hierarchy A <: B <: C and the abstracttype field is C, then both A and B are perfectly valid choices for subtype.)

abstracttype::Type: the abstract type under which the argument is meant to fall.

Constructors

NotImplementedError(::Function, ::Type, ::Type): constructs a new NotImplementedError instance for a single-argument function. Throws an error if the second type is not abstract or the first type is not a subtype of the second.

NotImplementedError(::Function, ::Symbol, ::Type, ::Type): constructs a new NotImplementedError instance for a multi-argument function. Throws an error if the second type is not abstract or the first type is not a subtype of the second.

MatrixBandwidth.RectangularMatrixError — Type

RectangularMatrixError(A)

TODO: Write here

MatrixBandwidth.StructuralAsymmetryError — Type

StructuralAsymmetryError(A, algorithm)

TODO: Write here. Cite [RS06].

MatrixBandwidth.Minimization.AbstractSolver — Type

AbstractSolver <: AbstractAlgorithm

Abstract base type for all matrix bandwidth minimization solvers.

Interface

As per the interface of supertype AbstractAlgorithm, concrete subtypes of AbstractSolver must implement the following methods:

Base.summary(::T) where {T<:AbstractSolver}: returns a String indicating the name of the solver (e.g., "Reverse Cuthill–McKee").

_requires_symmetry(::T) where {T<:AbstractSolver}: returns a Bool indicating whether the solver requires the input matrix to be structurally symmetric.

Direct subtypes of AbstractSolver must implement the following method:

_approach(::T) where {T<:AbstractSolver}: returns a Symbol indicating the category of solver (e.g., :heuristic).

MatrixBandwidth.Minimization.Exact.ExactSolver — Type

Exact <: AbstractSolver <: AbstractAlgorithm

Abstract type for all exact matrix bandwidth minimization solvers.

Exact methods are those which guarantee an optimal ordering producing the true minimum bandwidth of a matrix. Since bandwidth minimization is an NP-complete problem, existing exact algorithms are, at best, exponential in time complexity—much worse than many polynomial-time heuristic approaches (e.g., reverse Cuthill–McKee). Such methods, therefore, are not feasible for large matrices, but they remain useful when precise solutions are required for small-to-medium-sized inputs (say, up to $100×100$).

MatrixBandwidth.Minimization.Heuristic.HeuristicSolver — Type

HeuristicSolver <: AbstractSolver <: AbstractAlgorithm

Abstract type for all heuristic matrix bandwidth minimization solvers.

Heuristic methods are those which aim to produce near-optimal solutions in a more performant manner than exact methods. While precise bandwidth minimization is NP-complete, many heuristic algorithms (such as reverse Cuthill–McKee) run in polynomial time.

Heuristic algorithms differ from metaheuristic ones in that they do not employ higher-level iterative search frameworks (e.g., stochastic techniques) to survey the global search space and escape local minima; instead, they rely on straightforward deterministic procedures to find good solutions in a single pass.

MatrixBandwidth.Minimization.Heuristic.pseudo_peripheral_node — Method

pseudo_peripheral_node(A::AbstractMatrix{Bool}) -> Int

Select a pseudo-peripheral node from the connected graph represented by A.

This function acts as a node selector for the Cuthill–McKee and Reverse Cuthill–McKee algorithms, heuristically choosing the node "farthest" from the others in the graph. It is assumed that A is the adjacency matrix of some connected, undirected graph; otherwise, undefined behavior may arise.

Arguments

A::AbstractMatrix{Bool}: the adjacency matrix of some connected, undirected graph. In practice, this semantically represents the connected component of some larger graph.

Returns

Int: the index of the pseudo-peripheral node selected from the graph.

Notes

This function takes heavy inspiration from the implementation in [Net25], which accepts a graph object as input and leverages several pre-existing functions in the networkx library. We herein repurpose the logic to work directly on adjacency matrices, avoiding reallocation overhead and an unnecessary dependency on Graphs.jl.

MatrixBandwidth.Minimization.Metaheuristic.MetaheuristicSolver — Type

MetaheuristicSolver <: AbstractSolver <: AbstractAlgorithm

TODO: Write here

MatrixBandwidth.Recognition.AbstractDecider — Type

AbstractDecider <: AbstractAlgorithm

Abstract base type for all matrix bandwidth recognition deciders.

Interface

As per the interface of supertype AbstractAlgorithm, concrete subtypes of AbstractDecider must implement the following methods:

Base.summary(::T) where {T<:AbstractDecider}: returns a String indicating the name of the decider (e.g., "Caprara–Salazar-González").

_requires_symmetry(::T) where {T<:AbstractDecider}: returns a Bool indicating whether the decider requires the input matrix to be structurally symmetric.

References

[RS06]

J. K. Reid and J. A. Scott. Reducing the Total Bandwidth of a Sparse Unsymmetric Matrix. SIAM Journal on Matrix Analysis and Applications 28, 805–21 (2006).

[Net25]

NetworkX Developers. Source code for networkx.utils.rcm. NetworkX v3.5 documentation (2025). Accessed: 2025-06-11.