MatrixBandwidth.jl – Public API

Documentation for MatrixBandwidth's public API.

Note

The following documentation covers only the public API of the package. For internal details, see the private API documentation.

MatrixBandwidth.MatrixBandwidth — Module

MatrixBandwidth

Exact, heuristic, and metaheuristic algorithms for matrix bandwidth minimization in Julia.

The following algorithms are currently supported:

Exact

MBID: Minimum bandwidth by iterative deepening (MB-ID)

MBPS: Minimum bandwidth by perimeter search (MB-PS)

Heuristic

CuthillMcKee: Cuthill–McKee algorithm

ReverseCuthillMcKee: Reverse Cuthill–McKee algorithm

Metaheuristic

SimulatedAnnealing: Simulated annealing

GeneticAlgorithm: Genetic algorithm

GRASP: Greedy randomized adaptive search procedure (GRASP)

Full documentation is available for the latest development version of this package.

MatrixBandwidth.BandwidthResult — Type

BandwidthResult

TODO: Write here

MatrixBandwidth.bandwidth_unpermuted — Method

bandwidth_unpermuted(A) -> Int

TODO: Write here

MatrixBandwidth.minimize_bandwidth — Method

minimize_bandwidth(A, solver=ReverseCuthillMcKee()) -> BandwidthResult

TODO: Write here

MatrixBandwidth.Exact — Module

MatrixBandwidth.Exact

Exact solvers for matrix bandwidth minimization.

This submodule is part of the MatrixBandwidth.jl package.

MatrixBandwidth.Exact.MBID — Type

MBID <: ExactSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Exact.MBPS — Type

MBPS <: ExactSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Heuristic — Module

MatrixBandwidth.Heuristic

Heuristic solvers for matrix bandwidth minimization.

The following heuristic algorithms are currently supported:

CuthillMcKee: Cuthill–McKee algorithm

ReverseCuthillMcKee: Reverse Cuthill–McKee algorithm

This submodule is part of the MatrixBandwidth.jl package.

MatrixBandwidth.Heuristic.CuthillMcKee — Type

CuthillMcKee <: HeuristicSolver <: AbstractSolver

The Cuthill–McKee algorithm is a heuristic method for minimizing the bandwidth of a symmetric matrix $A$. It considers the graph $G(A)$ whose adjacency matrix is $A$ (ignoring self-loops) and performs a breadth-first search of each connected component of $G(A)$, starting from a low-degree node then visiting its neighbors in order of increasing degree. Particularly effective when $A$ is sparse, this heuristic typically produces an ordering which induces a matrix bandwidth either equal to or very close to the true minimum [CM69, pp. 157–58].

We also extend the algorithm to work more generally when $A$ is not symmetric by applying it to $A + Aᵀ$ instead, as suggested in [RS06, p. 808]. This approach still tends to produce a fairly good ordering, but it is not guaranteed to be as optimal as directly applying Cuthill–McKee to a symmetric input.

Fields

node_selector::Function: a function that selects a node from some connected component of the input matrix from which to start the breadth-first search. If no custom heuristic is specified, this field defaults to pseudo_peripheral_node, which picks a node "farthest" from the others in the component (not necessarily the lowest-degree node).

Examples

Cuthill–McKee finds an optimal ordering for an asymmetric $35×35$ matrix with bandwidth $3$ whose rows and columns have been shuffled:

julia> using Random

julia> Random.seed!(13);

julia> (n, k) = (35, 3);

julia> perm = randperm(n);

julia> A = MatrixBandwidth.random_sparse_banded_matrix(n, k);

julia> A_shuffled = A[perm, perm];

julia> res = minimize_bandwidth(A_shuffled, CuthillMcKee());

julia> iszero.(A_shuffled) == iszero.(A_shuffled') # Works even for asymmetric matrices
false

julia> bandwidth_unpermuted(A)
3

julia> bandwidth_unpermuted(A_shuffled)
33

julia> res.bandwidth # The true minimum bandwidth
3

Cuthill–McKee finds a near-optimal ordering for an asymmetric $200×200$ matrix with bandwidth $10$ whose rows and columns have been shuffled:

julia> using Random

julia> Random.seed!(37452);

julia> (n, k) = (200, 10);

julia> perm = randperm(n);

julia> A = MatrixBandwidth.random_sparse_banded_matrix(n, k);

julia> A_shuffled = A[perm, perm];

julia> res = minimize_bandwidth(A_shuffled, CuthillMcKee());

julia> iszero.(A_shuffled) == iszero.(A_shuffled') # Works even for asymmetric matrices
false

julia> bandwidth_unpermuted(A)
10

julia> bandwidth_unpermuted(A_shuffled)
194

julia> res.bandwidth # Close to the true minimum
14

Notes

It was found in [Geo71, pp. 114–15] that reversing the ordering produced by Cuthill–McKee tends to produce a better ordering; this variant is known as the reverse Cuthill–McKee algorithm. See ReverseCuthillMcKee and the associated method of _bool_minimal_band_ordering for our implementation of this algorithm.

Note also that the node_selector field must be of the form (A::AbstractMatrix{Bool}) -> Integer (i.e., it must take in an boolean matrix and return an integer). If this is not the case, an ArgumentError is thrown upon construction.

See also the documentation for supertypes HeuristicSolver and AbstractSolver.

MatrixBandwidth.Heuristic.ReverseCuthillMcKee — Type

ReverseCuthillMcKee <: HeuristicSolver <: AbstractSolver

The reverse Cuthill–McKee algorithm is a variant of the Cuthill–McKee algorithm—a heuristic method for minimizing the bandwidth of a symmetric matrix $A$. Cuthill–McKee considers the graph $G(A)$ whose adjacency matrix is $A$ (ignoring self-loops) and performs a breadth-first search of each connected component of $G(A)$, starting from a low-degree node then visiting its neighbors in order of increasing degree. Particularly effective when $A$ is sparse, this heuristic typically produces an ordering which induces a matrix bandwidth either equal to or very close to the true minimum [CM69, pp. 157–58]. The reverse Cuthill–McKee algorithm simply reverses the ordering produced by application of Cuthill–McKee; [Geo71, pp. 114–15] found that this tends to produce an even better ordering.

We also extend the algorithm to work more generally when $A$ is not symmetric by applying it to $A + Aᵀ$ instead, as suggested in [RS06, p. 808]. This approach still tends to produce a fairly good ordering, but it is not guaranteed to be as optimal as directly applying reverse Cuthill–McKee to a symmetric input.

Fields

node_selector::Function: a function that selects a node from some connected component of the input matrix from which to start the breadth-first search. If no custom heuristic is specified, this field defaults to pseudo_peripheral_node, which picks a node "farthest" from the others in the component (not necessarily the lowest-degree node).

Examples

Reverse Cuthill–McKee finds an optimal ordering for an asymmetric $45×45$ matrix with bandwidth $4$ whose rows and columns have been shuffled:

julia> using Random

julia> Random.seed!(87);

julia> (n, k) = (45, 4);

julia> perm = randperm(n);

julia> A = MatrixBandwidth.random_sparse_banded_matrix(n, k);

julia> A_shuffled = A[perm, perm];

julia> res = minimize_bandwidth(A_shuffled, ReverseCuthillMcKee());

julia> iszero.(A_shuffled) == iszero.(A_shuffled') # Works even for asymmetric matrices
false

julia> bandwidth_unpermuted(A)
4

julia> bandwidth_unpermuted(A_shuffled)
43

julia> res.bandwidth # The true minimum bandwidth
4

Reverse Cuthill–McKee finds a near-optimal ordering for an asymmetric $250×250$ matrix with bandwidth $14$ whose rows and columns have been shuffled:

julia> using Random

julia> Random.seed!(5747);

julia> (n, k) = (250, 14);

julia> perm = randperm(n);

julia> A = MatrixBandwidth.random_sparse_banded_matrix(n, k);

julia> A_shuffled = A[perm, perm];

julia> res = minimize_bandwidth(A_shuffled, ReverseCuthillMcKee());

julia> iszero.(A_shuffled) == iszero.(A_shuffled') # Works even for asymmetric matrices
false

julia> bandwidth_unpermuted(A)
14

julia> bandwidth_unpermuted(A_shuffled)
245

julia> res.bandwidth # Close to the true minimum
17

Notes

See CuthillMcKee and the associated method of _bool_minimal_band_ordering for our implementation of the original Cuthill–McKee algorithm. (Indeed, the reverse Cuthill–McKee method of _bool_minimal_band_ordering is merely a wrapper around the Cuthill–McKee method.)

Note also that the node_selector field must be of the form (A::AbstractMatrix{Bool}) -> Integer (i.e., it must take in an boolean matrix and return an integer). If this is not the case, an ArgumentError is thrown upon construction.

See also the documentation for supertypes HeuristicSolver and AbstractSolver.

MatrixBandwidth.Metaheuristic — Module

MatrixBandwidth.Metaheuristic

Metaheuristic solvers for matrix bandwidth minimization.

This submodule is part of the MatrixBandwidth.jl package.

MatrixBandwidth.Metaheuristic.GRASP — Type

GRASP <: MetaheuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Metaheuristic.GeneticAlgorithm — Type

GeneticAlgorithm <: MetaheuristicSolver <: AbstractSolver

TODO: Write here

MatrixBandwidth.Metaheuristic.SimulatedAnnealing — Type

SimulatedAnnealing <: MetaheuristicSolver <: AbstractSolver

TODO: Write here

References

[CM69]

E. Cuthill and J. McKee. Reducing the bandwidth of sparse symmetric matrices. In: Proceedings of the 24th National Conference of the ACM (Brandon Systems Press, 1969); pp. 157–72.

[Geo71]

J. A. George. Computer Implementation of the Finite Element Method. Ph.D. Thesis, Department of Computer Science, Stanford University (1971).

[RS06]

J. K. Reid and J. A. Scott. Reducing the Total Bandwidth of a Sparse Unsymmetric Matrix. SIAM Journal on Matrix Analysis and Applications 28, 805–21 (2006).